requirements/base.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.9.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.1
cbor2==5.7.1
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorlog==6.10.1
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==46.0.3
daphne==4.2.1
dj-database-url==3.0.1
Django==5.2.5
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-cuser==2017.3.16
django-debug-toolbar==6.0.0
django-filter==25.1
django-jazzmin==3.0.1
django-modeltranslation==0.19.16
django-query-counter==0.4.1
django-rosetta==0.10.2
django-timezone-field==7.1
django-tinymce==4.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
flower==2.0.1
frozenlist==1.8.0
gunicorn==23.0.0
humanize==4.12.3
hyperlink==21.0.0
idna==3.10
Incremental==24.11.0
inflection==0.5.1
iniconfig==2.1.0
kombu==5.5.4
langdetect==1.0.9
magic-filter==1.0.12
msgpack==1.1.2
multidict==6.7.0
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
polib==1.2.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.4.1
psycopg2-binary==2.9.10
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pytest==8.4.1
pytest-django==4.11.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
requests==2.32.4
ruff==0.12.9
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
tabulate==0.9.0
tornado==6.5.2
Twisted==25.5.0
txaio==25.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.22.0
zope.interface==8.1.1
